feat: updated Linters github action
