+++ .github/workflows/linters.yml
@@ -13,10 +13,12 @@ jobs:
 
     steps:
       - name: Checkout code
-        uses: actions/checkout@v4
+        uses: actions/checkout@v6
+        with:
+          fetch-depth: 0
 
       - name: Setup Node.js
-        uses: actions/setup-node@v4
+        uses: actions/setup-node@v6
         with:
           node-version: 22
           cache: npm
@@ -32,6 +34,11 @@ jobs:
 
       - name: Run type verification
         run: npm run verify
+
+      - name: Run Fallow quality gate
+        env:
+          FALLOW_BASE: ${{ github.event_name == 'pull_request' && github.event.pull_request.base.sha || 'origin/master' }}
+        run: npm run fallow:audit:ci -- --base "$FALLOW_BASE"
 
       - name: Run tests
         run: npm test
