+++ src/index.ts
@@ -1,9 +1,5 @@
 import BoundlessCheckout from "./BoundlessCheckout";
-import {startCheckout, StarterWrapper, resetCheckoutState} from "./starters";
 
 export {
 	BoundlessCheckout,
-	startCheckout,
-	StarterWrapper,
-	resetCheckoutState
 };
