+++ src/pages/paymentPage/types.ts
@@ -1,0 +1,8 @@
+export interface IPaymentMethod {
+	payment_method_id: string;
+	title: string;
+}
+
+export interface IPaymentPageData {
+	paymentMethods: IPaymentMethod[];
+}
