+++ src/pages/shippingPage/shippingForm/AddressFieldset.tsx
@@ -120,20 +120,7 @@ interface IProps {
 	keyPrefix: "shipping_address" | "billing_address";
 }
 
-export interface IAddressFields {
-	first_name?: string;
-	last_name?: string;
-	company?: string;
-	address_line_1?: string;
-	address_line_2?: string;
-	city?: string;
-	state?: string;
-	country_id?: number | string;
-	zip?: string;
-	phone?: string;
-}
-
-export function addressFieldAttrs(
+function addressFieldAttrs(
 	keyPrefix: "shipping_address" | "billing_address",
 	field: string,
 	formikProps: FormikProps<IShippingFormValues>,
