--- src/redux/actions/app.ts
@@ -1,8 +1,0 @@
-import {AppThunk} from "../store";
-import {resetAppState} from "../reducers/app";
-import {resetXhrState} from "../reducers/xhr";
-
-export const resetState = () : AppThunk => async (dispatch) => {
-	dispatch(resetAppState());
-	dispatch(resetXhrState());
-};
